+++ packages/appcd-plugin/CHANGELOG.md
@@ -8,11 +8,13 @@
    [(DAEMON-314)](https://jira.appcelerator.org/browse/DAEMON-314)
  * feat: Added `description`, `homepage`, `license`, `appcdVersion`, `apiVersion`, and `os` from
    the plugin's `package.json` to the plugin info.
+ * feat: Added plugin start time to the plugin info.
  * feat(scheme): Added `detect()` method to schemes that resolves a list of plugins found. Also
    managed to clean up a bunch of code and tests in the process.
  * feat: Added symlink flag to plugin descriptor. Used to guess if plugin is a yarn link.
  * feat: Added support for a `autoStart` flag in the plugin's `package.json`.
  * feat: Redact specific sensitive data from plugin debug logging.
+ * feat: Added `appcd.logger` for plugins to create new namespaced loggers.
  * fix: Fixed bug where `status` would search an unsorted list of registered plugins which caused
    nondeterministic results. [(DAEMON-328)](https://jira.appcelerator.org/browse/DAEMON-328)
  * fix: Fixed bug where an `inactivityTimeout` of zero would be tested as falsey and fallback to